Move appconfig.php to config dir
