+++ src/views/home.php
@@ -1,4 +1,4 @@
-<?php require_once('./resources/appconfig.php') ?>
+<?php require_once('config/appconfig.php') ?>
 <!DOCTYPE html>
 <html lang="en">
